# api.yaml
openapi: 3.0.3
info:
  title: Dorra EMR API
  version: 1.0.0 (v1)
  description: |-
    API documentation for Dorra EMR

    View Swagger docs [here](https://hackathon-api.aheadafrica.org/docs/swagger)
paths:
...
  /v1/appointments:
    get:
      operationId: appointments_list
      description: List all appointments created in your team
      summary: List all appointments
      parameters:
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - appointments
      security:
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAppointmentList'
          description: ''
  /v1/appointments/{id}:
    get:
      operationId: appointments_retrieve
      description: Retrieve appointment details
      summary: Retrieve an appointment
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - appointments
      security:
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
          description: ''
    patch:
      operationId: appointments_partial_update
      description: Update appointment details. This endpoint supports partial updates,
        meaning that only the fields to be modified need to be included in the request.
      summary: Update an appointment
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - appointments
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedAppointment'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedAppointment'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedAppointment'
      security:
      - tokenAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdatedResourceResponse'
          description: ''
    delete:
      operationId: appointments_destroy
      description: Delete an appointment
      summary: Delete a patient
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - appointments
      security:
      - tokenAuth: []
      responses:
        '204':
          description: No response body
...
  /v1/patients/{id}:
    get:
      operationId: patients_retrieve
      description: Retrieve a patient in your team
      summary: Retrieve a patient
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - patients
      security:
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
          description: ''
    patch:
      operationId: patients_partial_update
      description: Update patient details. This endpoint allows partial updates, so
        only the fields that require modification need to be included in the request.
      summary: Update a patient
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - patients
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedPatient'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedPatient'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedPatient'
      security:
      - tokenAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdatedResourceResponse'
          description: ''
    delete:
      operationId: patients_destroy
      description: Delete a patient
      summary: Delete a patient
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - patients
      security:
      - tokenAuth: []
      responses:
        '204':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletedResourceResponse'
          description: ''
  /v1/patients/{id}/appointments:
    get:
      operationId: patients_appointments_list
      description: List all appointments for a patient
      summary: List patient appointments
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - patients
      security:
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAppointmentList'
          description: ''
...
  /v1/patients/create:
    post:
      operationId: patients_create_create
      description: Create a patient via standard endpoint
      summary: Create Patient
      tags:
      - patients
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatient'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CreatePatient'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreatePatient'
        required: true
      security:
      - tokenAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedPatientResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthenticatedResponse'
          description: ''
...
components:
  schemas:
    Appointment:
      type: object
      description: encounter serializer
      properties:
        id:
          type: integer
          readOnly: true
        patient_name:
          type: string
          readOnly: true
        date:
          type: string
          format: date-time
        reason:
          type: string
          nullable: true
        summary:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/StatusEnum'
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        patient:
          type: integer
          nullable: true
      required:
      - created_at
      - id
      - patient_name
      - updated_at
...
    BadRequestResponse:
      type: object
      properties:
        status:
          type: boolean
          default: false
        status_code:
          type: integer
          default: 400
        message:
          type: string
          default: Invalid request. please check the prompt and try again.
...
    CreatePatient:
      type: object
      description: Patient serializer
      properties:
        id:
          type: integer
          readOnly: true
        allergies:
          type: array
          items:
            type: string
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          nullable: true
          maxLength: 100
        age:
          type: string
          nullable: true
        date_of_birth:
          type: string
          format: date
          nullable: true
        gender:
          nullable: true
          oneOf:
          - $ref: '#/components/schemas/GenderEnum'
          - $ref: '#/components/schemas/NullEnum'
        address:
          type: string
          nullable: true
        phone_number:
          type: string
          nullable: true
          maxLength: 20
        email:
          type: string
          format: email
          nullable: true
          maxLength: 254
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - first_name
      - id
      - updated_at
...
    CreatedPatientResponse:
      type: object
      properties:
        status:
          type: boolean
          default: true
        status_code:
          type: integer
          default: 201
        message:
          type: string
          default: success
        id:
          type: integer
          default: 67
    DeletedResourceResponse:
      type: object
      properties:
        status:
          type: boolean
          default: true
        status_code:
          type: integer
          default: 204
        message:
          type: string
          default: success
        id:
          type: integer
          default: 67
...
    GenderEnum:
      enum:
      - Male
      - Female
      - Other
      type: string
      description: |-
        * `Male` - Male
        * `Female` - Female
        * `Other` - Other
...
    NullEnum:
      enum:
      - null
    PaginatedAppointmentList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Appointment'
...
    PatchedAppointment:
      type: object
      description: encounter serializer
      properties:
        id:
          type: integer
          readOnly: true
        patient_name:
          type: string
          readOnly: true
        date:
          type: string
          format: date-time
        reason:
          type: string
          nullable: true
        summary:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/StatusEnum'
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        patient:
          type: integer
          nullable: true
    PatchedPatient:
      type: object
      description: Patient serializer
      properties:
        id:
          type: integer
          readOnly: true
        unique_id:
          type: string
          maxLength: 10
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          nullable: true
          maxLength: 100
        full_name:
          type: string
          nullable: true
          maxLength: 200
        age:
          type: string
          nullable: true
        date_of_birth:
          type: string
          format: date
          nullable: true
        gender:
          nullable: true
          oneOf:
          - $ref: '#/components/schemas/GenderEnum'
          - $ref: '#/components/schemas/NullEnum'
        address:
          type: string
          nullable: true
        phone_number:
          type: string
          nullable: true
          maxLength: 20
        email:
          type: string
          format: email
          nullable: true
          maxLength: 254
        allergies: {}
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        team:
          type: integer
          nullable: true
    Patient:
      type: object
      description: Patient serializer
      properties:
        id:
          type: integer
          readOnly: true
        unique_id:
          type: string
          maxLength: 10
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          nullable: true
          maxLength: 100
        full_name:
          type: string
          nullable: true
          maxLength: 200
        age:
          type: string
          nullable: true
        date_of_birth:
          type: string
          format: date
          nullable: true
        gender:
          nullable: true
          oneOf:
          - $ref: '#/components/schemas/GenderEnum'
          - $ref: '#/components/schemas/NullEnum'
        address:
          type: string
          nullable: true
        phone_number:
          type: string
          nullable: true
          maxLength: 20
        email:
          type: string
          format: email
          nullable: true
          maxLength: 254
        allergies: {}
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        team:
          type: integer
          nullable: true
      required:
      - created_at
      - first_name
      - id
      - unique_id
      - updated_at
...
    StatusEnum:
      enum:
      - active
      - completed
      type: string
      description: |-
        * `active` - Active
        * `completed` - Completed
...
    UnauthenticatedResponse:
      type: object
      properties:
        status:
          type: boolean
          default: false
        status_code:
          type: integer
          default: 401
        message:
          type: string
        details:
          type: string
      required:
      - details
      - message
    UpdatedResourceResponse:
      type: object
      properties:
        status:
          type: boolean
          default: true
        status_code:
          type: integer
          default: 200
        message:
          type: string
          default: success
        id:
          type: integer
          default: 67
...
  securitySchemes:
    tokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Token-based authentication with required prefix "Token"